refactor: renaming marshaling/unmarshaling files and functions for both json and pongo schema
* renamed schema_marshal.go to schema_pongoschema_marshal.go
* removed schema_interface.go, split across schema.go, schema_jsonschema_marshal.go, schema_pongoschema_marshal.go
* renamed from DefaultSchemaUnmarshalMap to GlobalSchemaUnmarshalMap
