+++ README.md
@@ -255,7 +255,7 @@ schema = pongo.Object(pongo.O{
     "aInt":    pongo.Int(),
     "aString": pongo.String(),
 })
-marshalledSchema, err := pongo.MarshalSchemaJSON(schema)
+marshalledSchema, err := pongo.MarshalPongoSchema(schema)
 if err != nil {
     panic("unexpected error during marshalling: " + err.Error())
 } else {
@@ -272,7 +272,7 @@ marshalled schema:
 To unmarshall back the schema
 
 ```go
-newSchema, _, err := pongo.UnmarshalSchemaJSON(marshalledSchema)
+newSchema, _, err := pongo.UnmarshalPongoSchema(marshalledSchema)
 if err != nil {
     panic("unexpected error during unmarshalling: " + err.Error())
 } else {
@@ -303,6 +303,6 @@ If you need to unmarshal your custom type, you'll also need to use a custom `Sch
 When you'll unmarshal your PonGO schema, you must pass to the unmarshal function your custom map
 
 ```go
-myCustomMapper = pongo.DefaultSchemaUnmarshalMap().set()
-myPonGOSchema, err := pongo.UnmarshalSchemaJSONWithMapper(myPonGOSchemaJSON, myCustomMapper)
+myCustomMapper = pongo.GlobalSchemaUnmarshalMap().set()
+myPonGOSchema, err := pongo.UnmarshalSchemaWithMapper(myPonGOSchemaJSON, myCustomMapper)
 ```
